Added typeclasses to inferencer
Hopefully they work as intended. There's currently one class, Concrete, which holds non-function types and supports comparison, truthiness and string conversion. Builtins have typeclass constraints, and if the -i flag is not given, the inferencer will constrain the result type to be an instance of Concrete.

In the code, custom typeclasses can be added by extending the ADT, the Types instance and the "holds" function, which checks whether a typeclass constraint holds.